+++ src/components/content/patient-history/use-builder-patient-history-list.ts
@@ -1,0 +1,52 @@
+import { useQueryWithPatient } from "@zus-health/ctw-component-library";
+import {
+  getBuilderRefreshHistoryMessages,
+  PatientHistoryResponse,
+} from "./use-patient-history";
+import { PatientHistoryPatient } from "@/fhir/models/patient-history";
+import { getBuilderPatientsListByIdentifier } from "@/fhir/patient-helper";
+import { compact } from "@/utils/nodash";
+import { QUERY_KEY_PATIENT_HISTORY_LIST } from "@/utils/query-keys";
+import { Telemetry } from "@/utils/telemetry";
+
+export function useBuilderPatientHistoryList(
+  pageSize: number,
+  pageOffset: number
+) {
+  return useQueryWithPatient(
+    QUERY_KEY_PATIENT_HISTORY_LIST,
+    [pageSize, pageOffset],
+    async (requestContext) => {
+      try {
+        const messages = (
+          await getBuilderRefreshHistoryMessages(requestContext)
+        ).data as PatientHistoryResponse[];
+
+        const patientsIds = compact(
+          messages.map((message) => message.initialData?.patientId)
+        );
+
+        const patientData = await getBuilderPatientsListByIdentifier(
+          requestContext,
+          undefined,
+          patientsIds
+        );
+
+        const patientHistoryPatients = patientData.patients.map((patient) => {
+          const matchingPatientId = messages.filter(
+            (message) => message.initialData?.patientId === patient.id
+          );
+          return new PatientHistoryPatient(patient, matchingPatientId[0]);
+        });
+
+        return { total: patientData.total, patients: patientHistoryPatients };
+      } catch (e) {
+        Telemetry.logError(
+          e as Error,
+          "Failed fetching patient history patients."
+        );
+        throw new Error(`Failed fetching patient history patients: ${e}`);
+      }
+    }
+  );
+}
